refactor filter_resources to base class for filter/actions to remove dupe
