+++ c7n/element.py
@@ -1,0 +1,40 @@
+# Copyright 2020 Cloud Custodian Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import jmespath
+
+
+class Element:
+    """Parent base class for filters and actions.
+    """
+
+    def filter_resources(self, resources, key_expr, allowed_values=()):
+        # many filters implementing a resource state transition only allow
+        # a given set of starting states, this method will filter resources
+        # and issue a warning log, as implicit filtering in filters means
+        # our policy metrics are off, and they should be added as policy
+        # filters.
+        resource_count = len(resources)
+        search_expr = key_expr
+        if not search_expr.startswith('[].'):
+            search_expr = '[].' + key_expr
+        results = [r for value, r in zip(
+            jmespath.search(search_expr, resources), resources)
+            if value in allowed_values]
+        if resource_count != len(results):
+            self.log.warning(
+                "%s implicitly filtered %d of %d resources key:%s on %s",
+                self.type, len(results), resource_count, key_expr,
+                (', '.join(allowed_values)))
+        return results
