+++ c7n/filters/core.py
@@ -30,6 +30,7 @@ from dateutil.parser import parse
 from distutils import version
 import jmespath
 
+from c7n.element import Element
 from c7n.exceptions import PolicyValidationError
 from c7n.executor import ThreadPoolExecutor
 from c7n.registry import PluginRegistry
@@ -167,7 +168,7 @@ class FilterRegistry(PluginRegistry):
 # Really should be an abstract base class (abc) or
 # zope.interface
 
-class Filter:
+class Filter(Element):
 
     executor_factory = ThreadPoolExecutor
 
@@ -218,26 +219,6 @@ class Filter:
 
         if not values and block_op != 'or':
             return
-
-    def filter_resources(self, resources, key_expr, allowed_values=()):
-        # many filters implementing a resource state transition only allow
-        # a given set of starting states, this method will filter resources
-        # and issue a warning log, as implicit filtering in filters means
-        # our policy metrics are off, and they should be added as policy
-        # filters.
-        resource_count = len(resources)
-        search_expr = key_expr
-        if not search_expr.startswith('[].'):
-            search_expr = '[].' + key_expr
-        results = [r for value, r in zip(
-            jmespath.search(search_expr, resources), resources)
-            if value in allowed_values]
-        if resource_count != len(results):
-            self.log.warning(
-                "%s implicitly filtered %d of %d resources key:%s on %s",
-                self.type, len(results), resource_count, key_expr,
-                (', '.join(allowed_values)))
-        return results
 
 
 def intersect_list(a, b):
